requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.1.0
numpy==1.24.3
tensorflow==2.15.0
scikit-learn==1.3.2
xgboost==2.0.3
pydantic==2.5.0
